# services/user-service/docs/swagger.yaml
definitions:
...
  github_com_Segun228_MAN_Alpha_bot_services_user-service_internal_models.Report:
    properties:
      agr:
        type: number
      apc:
        type: integer
      avp:
        type: number
      cogs:
        type: number
      cogs1s:
        type: number
      created_at:
        type: string
      customers:
        type: integer
      fc:
        type: number
      id:
        type: integer
      name:
        type: string
      rr:
        type: number
      tms:
        type: number
      updated_at:
        type: string
      user_id:
        type: integer
      users:
        type: integer
    type: object
...
  internal_controller_http_v1.reportCreateRequest:
    properties:
      agr:
        type: number
      apc:
        type: integer
      avp:
        type: number
      cogs:
        type: number
      cogs1s:
        type: number
      customers:
        type: integer
      fc:
        type: number
      name:
        type: string
      rr:
        type: number
      tms:
        type: number
      user_id:
        type: integer
      users:
        type: integer
    type: object
  internal_controller_http_v1.reportUpdateRequest:
    properties:
      agr:
        type: number
      apc:
        type: integer
      avp:
        type: number
      cogs:
        type: number
      cogs1s:
        type: number
      customers:
        type: integer
      fc:
        type: number
      name:
        type: string
      rr:
        type: number
      tms:
        type: number
      users:
        type: integer
    type: object
...
info:
  contact: {}
  title: User Service
  version: "1.0"
paths:
...
  /reports:
    get:
      consumes:
      - application/json
      description: Получить список всех отчетов
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_Segun228_MAN_Alpha_bot_services_user-service_internal_models.Report'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all reports
      tags:
      - reports
    post:
      consumes:
      - application/json
      description: Создать новый отчет
      parameters:
      - description: Report to create
        in: body
        name: report
        required: true
        schema:
          $ref: '#/definitions/internal_controller_http_v1.reportCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Segun228_MAN_Alpha_bot_services_user-service_internal_models.Report'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new report
      tags:
      - reports
  /reports/{reportID}:
    delete:
      consumes:
      - application/json
      description: Удалить отчет по ID
      parameters:
      - description: ID отчета
        in: path
        name: reportID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete report by ID
      tags:
      - reports
    patch:
      consumes:
      - application/json
      description: Обновить отчет по ID
      parameters:
      - description: ID отчета
        in: path
        name: reportID
        required: true
        type: integer
      - description: Report fields to update
        in: body
        name: report
        required: true
        schema:
          $ref: '#/definitions/internal_controller_http_v1.reportUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Segun228_MAN_Alpha_bot_services_user-service_internal_models.Report'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update report by ID (full)
      tags:
      - reports
    put:
      consumes:
      - application/json
      description: Частичное обновление отчета по ID
      parameters:
      - description: ID отчета
        in: path
        name: reportID
        required: true
        type: integer
      - description: Report fields to update
        in: body
        name: report
        required: true
        schema:
          $ref: '#/definitions/internal_controller_http_v1.reportUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Segun228_MAN_Alpha_bot_services_user-service_internal_models.Report'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Patch report by ID (partial)
      tags:
      - reports
  /reports/tg/{tgID}:
    get:
      consumes:
      - application/json
      description: Получить все отчеты пользователя по Telegram ID
      parameters:
      - description: Telegram User ID
        format: int64
        in: path
        name: tgID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_Segun228_MAN_Alpha_bot_services_user-service_internal_models.Report'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get reports by user Telegram ID
      tags:
      - reports
    post:
      consumes:
      - application/json
      description: Создать новый отчет, используя Telegram ID пользователя
      parameters:
      - description: Telegram User ID
        format: int64
        in: path
        name: tgID
        required: true
        type: integer
      - description: Report to create
        in: body
        name: report
        required: true
        schema:
          $ref: '#/definitions/internal_controller_http_v1.reportCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Segun228_MAN_Alpha_bot_services_user-service_internal_models.Report'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new report with Telegram ID
      tags:
      - reports
...
swagger: "2.0"